FlaskService/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
azure-ai-documentintelligence==1.0.2
azure-ai-formrecognizer==3.3.3
azure-cognitiveservices-speech==1.47.0
azure-common==1.1.28
azure-core==1.36.0
azure-functions==1.24.0
azure-identity==1.25.1
azure-search-documents==11.6.0
azure-storage-blob==12.27.1
blinker==1.9.0
cachetools==6.2.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
distro==1.9.0
dotenv==0.9.9
Flask==3.0.2
Flask-Cors==4.0.0
google-auth==2.43.0
google-genai==1.52.0
h11==0.16.0
holidays==0.36
httpcore==1.0.9
httpx==0.28.1
idna==3.11
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
kaleido==0.2.1
MarkupSafe==3.0.3
msal==1.34.0
msal-extensions==1.3.1
msrest==0.7.1
numpy==2.3.4
oauthlib==3.3.1
openai==2.7.2
packaging==25.0
plotly==5.22.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
python-dateutil==2.9.0
python-dotenv==1.2.1
requests==2.31.0
requests-oauthlib==2.0.0
rerequests==1.0.0b0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.1.0
websockets==15.0.1
Werkzeug==3.1.3

